feat: add support for sudo path in Global class
This commit adds a new private variable `_sudoPath` to the `Global` class in `lib/global.dart`. It also adds a getter and setter for this variable. The `sudoPath` getter returns the value of `_sudoPath`, and the `sudoPath` setter sets the value of `_sudoPath`.

Additionally, this commit modifies `lib/os/prepare.dart` to use the `Global.sudoPath` variable instead of calling `Executable('sudo').find()`. This allows the `sudo` path to be set globally and used throughout the application.