feat(shortcut): add support for removing shortcuts
This commit adds a new flag to the `ShortcutCommand` class, `--remove` or `-r`, which allows the user to remove a shortcut. If this flag is present, the `Shortcut` class will call the `_delete` method instead of the `_create` method.

Additionally, the `Shortcut` class now has a `_delete` method that deletes the shortcut. This method is implemented for Linux-based systems and Windows, but not for macOS.

Finally, the `Shortcut` class now imports the `path` package, which was previously missing.